Refactor Borrow hooks
